timeout fix

+++ tests/test_util.py
@@ -286,7 +286,7 @@ def test_split_files():
     file_size = 0.1  # MB
     num_files = 3
     chunk_size = 0.01  # MB
-    files = {"file_%s.txt" % i: {"content": rand_string(file_size * 1024 * 1024), "offset": 0}  for i in range(num_files)}
+    files = {"file_%s.txt" % i: {"content": rand_string(int(file_size * 1024 * 1024)), "offset": 0}  for i in range(num_files)}
     chunks = util.split_files(files, MAX_MB=chunk_size)
 
     # re combine chunks
